requirements.txt
# used Python 3.10
matplotlib~=3.8.2
numpy~=1.26.3
pygame-ce~=2.4.0
scikit-fuzzy latest
pygame-matplotlib~=0.4
